Move SystemErrorAlert to separate folder
