+++ apps/melding-form/src/app/(general)/Home.tsx
@@ -10,7 +10,7 @@ import { InvalidFormAlert } from '@meldingen/ui'
 
 import type { FormState } from '~/types'
 
-import { SystemErrorAlert } from './_components/SystemErrorAlert'
+import { SystemErrorAlert } from './_components'
 import { getDocumentTitleOnError } from './_utils/validation/getDocumentTitleOnError'
 
 const initialState: FormState = {}
