+++ apps/melding-form/src/app/(general)/_components/index.ts
@@ -1,2 +1,3 @@
 export { BackOfficeLayout } from './BackOfficeLayout/BackOfficeLayout'
 export { RegularLayout } from './RegularLayout/RegularLayout'
+export { SystemErrorAlert } from './SystemErrorAlert/SystemErrorAlert'
