+++ apps/melding-form/src/app/(general)/samenvatting/Summary.tsx
@@ -10,7 +10,7 @@ import { Link, SubmitButton, SummaryList, UnorderedList } from '@meldingen/ui'
 
 import type { FormState } from '~/types'
 
-import { SystemErrorAlert } from '../_components/SystemErrorAlert'
+import { SystemErrorAlert } from '../_components'
 import { getDocumentTitleOnError } from '../_utils/validation/getDocumentTitleOnError'
 import { BackLink } from '../../_components'
 import { AttachmentImage } from './_components/AttachmentImage'
